feat: chat list for spaces Show the regular chat list for spaces and move the UI to join spaces to a separate screen.
