+++ lib/config/routes.dart
@@ -33,6 +33,7 @@ import 'package:fluffychat/pages/settings_password/settings_password.dart';
 import 'package:fluffychat/pages/settings_security/settings_security.dart';
 import 'package:fluffychat/pages/settings_style/settings_style.dart';
 import 'package:fluffychat/pages/sign_in/sign_in_page.dart';
+import 'package:fluffychat/pages/space_rooms/space_rooms.dart';
 import 'package:fluffychat/widgets/config_viewer.dart';
 import 'package:fluffychat/widgets/layouts/empty_page.dart';
 import 'package:fluffychat/widgets/layouts/two_column_layout.dart';
@@ -422,6 +423,15 @@ abstract class AppRoutes {
                     InvitationSelection(
                       roomId: state.pathParameters['roomid']!,
                     ),
+                  ),
+                  redirect: loggedOutRedirect,
+                ),
+                GoRoute(
+                  path: 'rooms',
+                  pageBuilder: (context, state) => defaultPageBuilder(
+                    context,
+                    state,
+                    SpaceRoomsPage(spaceId: state.pathParameters['roomid']!),
                   ),
                   redirect: loggedOutRedirect,
                 ),
