+++ lib/l10n/intl_en.arb
@@ -2551,6 +2551,7 @@
         }
     },
     "deviceIdentityKey": "Device identity key",
+    "rooms": "Rooms",
     "restoreBootstrapEmptyDevicesDescription": "Please enter your passphrase or recovery key to verify this device and get access to your encrypted message backup.",
     "@restoreBootstrapEmptyDevicesDescription": {
         "type": "String",
