+++ lib/pages/chat_list/chat_list_item.dart
@@ -22,6 +22,7 @@ class ChatListItem extends StatelessWidget {
   final Room room;
   final Room? space;
   final bool activeChat;
+  final String? activeSpace;
   final void Function(BuildContext context)? onLongPress;
   final void Function()? onForget;
   final void Function() onTap;
@@ -30,6 +31,7 @@ class ChatListItem extends StatelessWidget {
   const ChatListItem(
     this.room, {
     this.activeChat = false,
+    this.activeSpace,
     required this.onTap,
     this.onLongPress,
     this.onForget,
@@ -88,7 +90,7 @@ class ChatListItem extends StatelessWidget {
                     height: Avatar.defaultSize,
                     child: Stack(
                       children: [
-                        if (space != null)
+                        if (space != null && space.id != activeSpace)
                           Positioned(
                             top: 0,
                             left: 0,
